requirements.txt
albumentations==1.4.15
matplotlib==3.9.2
numpy~=1.26.4
pandas==2.2.2
Pillow==10.4.0
pynrrd==1.0.0
pytorch_lightning==2.4.0
scikit_learn==1.5.2
segmentation_models_pytorch==0.3.4
tabulate==0.9.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5

pytorch-lightning~=2.4.0
lightning~=2.4.0
segmentation-models-pytorch~=0.3.4
wandb~=0.18.1
torchmetrics~=1.4.2